WORK: 1. ivy 完善，做成独立库，参考 linux kernel 6.6.31 调整源码结构，复用 linux kernel 的类型、原子操作等，实现可重复使用的 barrier, spinlock 改用 linux kernel 的排队锁

+++ memtester/memtester.h
@@ -16,9 +16,10 @@
 // Copyright (C) 2024 zuoqian
 
 // #include <sys/types.h>
+#include <stdint.h>
 
 /* extern declarations. */
-typedef long off_t;
+typedef ptrdiff_t off_t;
 
 extern int use_phys;
 extern off_t physaddrbase;
